Windows: Fix build and start automated testing
* Remove `--timeout` on Windows since it relies on Unix-only code (#74).
* Update PR checks to run on Windows.

Fixes: #73 — Test on Windows
